test: expanded tests
