+++ conftest.py
@@ -1,13 +1,8 @@
 import asyncio
-from asyncio import AbstractEventLoop
-from collections.abc import Generator
 
 import pytest
 
 
-@pytest.fixture(scope="session", autouse=True)
-def event_loop() -> Generator[AbstractEventLoop, None, None]:
-    """Create an instance of the default event loop for the test session."""
-    loop = asyncio.new_event_loop()
-    yield loop
-    loop.close()
+@pytest.fixture(scope="session")
+def event_loop_policy():
+    return asyncio.get_event_loop_policy()
